requirements.txt

discord.py>=2.3.2
flask
psycopg2-binary
python-dotenv
